feat: add some checks in uuid and optional parameters

+++ src/app/controllers/ContactController.js
@@ -33,14 +33,22 @@ class ContactController {
       return response.status(400).json({ error: 'name is required' });
     }
 
-    const contactExists = await ContactRepository.findByEmail(email);
+    if (category_id && !isValidUUID(category_id)) {
+      return response.status(400).json({ error: 'invalid category' });
+    }
 
-    if (contactExists) {
-      return response.status(400).json({ error: 'this e-mail is already in use' });
+    if (email) {
+      const contactExists = await ContactRepository.findByEmail(email);
+      if (contactExists) {
+        return response.status(400).json({ error: 'this e-mail is already in use' });
+      }
     }
 
     const contact = await ContactRepository.create({
-      name, email, phone, category_id,
+      name,
+      email: email || null,
+      phone,
+      category_id: category_id || null,
     });
 
     response.status(201).json(contact);
@@ -56,22 +64,31 @@ class ContactController {
       return response.status(400).json({ error: 'invalid user id' });
     }
 
+    if (!name) {
+      return response.status(400).json({ error: 'name is required' });
+    }
+
+    if (category_id && !isValidUUID(category_id)) {
+      return response.status(400).json({ error: 'invalid category' });
+    }
+
     const contactExists = await ContactRepository.findById(id);
     if (!contactExists) {
       return response.status(404).json({ error: 'contact not found' });
     }
 
-    if (!name) {
-      return response.status(400).json({ error: 'name is required' });
-    }
-
-    const contactByEmail = await ContactRepository.findByEmail(email);
-    if (contactByEmail && contactByEmail.id !== id) {
-      return response.status(400).json({ error: 'this e-mail is already in use' });
+    if (email) {
+      const contactByEmail = await ContactRepository.findByEmail(email);
+      if (contactByEmail && contactByEmail.id !== id) {
+        return response.status(400).json({ error: 'this e-mail is already in use' });
+      }
     }
 
     const contact = await ContactRepository.update(id, {
-      name, email, phone, category_id,
+      name,
+      email: email || null,
+      phone,
+      category_id: category_id || null,
     });
 
     response.json(contact);
@@ -79,6 +96,9 @@ class ContactController {
 
   async delete(request, response) {
     const { id } = request.params;
+    if (!isValidUUID(id)) {
+      return response.status(400).json({ error: 'invalid user id' });
+    }
 
     await ContactRepository.delete(id);
